Added function to get associated widgets from datasetId
